Progress on rent but now need to debit it from cash and credit to tocks when paid etc

+++ hive/h/types.h
@@ -20,7 +20,7 @@
 typedef uint64_t Nanosecs;
 typedef uint32_t Tocks;  // A tock might be around a microsecond, but varies
 typedef uint32_t Index;
-typedef uint32_t TockDuration; // In nanoseconds.
+typedef uint64_t TockDuration; // In nanoseconds.
 typedef uint64_t Score;
 typedef uint64_t Cash;  
 typedef uint32_t TockPrice; // The currency unit can be tiny.
